Datagen

+++ noise.py
@@ -10,9 +10,8 @@ def add_noise(X, alpha=0.1):
 def generate_noise_seqs(X):
     seqs = X.clone()[:, None, :]
 
-    while torch.isclose(X, torch.tensor([2], dtype=float)).sum() != X.size(dim=0) * X.size(dim=1):
+    while torch.isclose(X, torch.tensor([2], dtype=torch.float32)).sum() != X.size(dim=0) * X.size(dim=1):
         X = add_noise(X)
 
         seqs = torch.cat((seqs, X[:, None, :]), 1)
-
     return seqs
